pyproject.toml
[project]
name = "hp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.3.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "channels[daphne]>=4.3.2",
    "channels-redis>=4.3.0",
]
